Fixes build and closes #2
Builds were failing due to the use of the `autoM` function from
`Data.Machine`. This function is from `Data.Machine.Process` and is only
exported on `machines >=0.2.4`. This changes the lower bound of the
`machines` dependency to reflect that and fixes the build. It closes the
issue #2 and should also close issue aloiscochard/codex#24 only
published and used there.

For reference:
- [`Data.Machine.Process@0.2.3.1`](http://hackage.haskell.org/package/machines-0.2.3.1/docs/src/Data-Machine-Process.html)
- [`Data.Machine.Process@0.2.4`](http://hackage.haskell.org/package/machines-0.2.4/docs/src/Data-Machine-Process.html)